Final/requirements.txt
appnope==0.1.0
appscript==1.0.1
boto==2.49.0
click==8.0.1
entrypoints==0.3
Flask==2.0.1
gmpy2==2.0.8
importlib-metadata==4.6.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
locket==0.2.0
MarkupSafe==2.0.1
mpmath==1.1.0
psycopg2==2.9.1
ptyprocess==0.6.0
pycosat==0.6.3
pycrypto==2.6.1
pycurl==7.43.0.5
pypng==0.0.20
PyQRCode==1.2.1
PyYAML==5.3
qrcode==7.2
termcolor==1.1.0
terminado==0.8.3
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0